Refactor storage utils ♻️
