+++ server/shared/storage/storage.controller.js
@@ -1,9 +1,8 @@
 'use strict';
 
 const { getFileUrl, saveFile } = require('./');
+const { readFile, sha256 } = require('./util');
 const { ASSET_ROOT } = require('./helpers');
-const crypto = require('crypto');
-const fs = require('fs');
 const path = require('path');
 
 function getUrl(req, res) {
@@ -24,13 +23,4 @@ async function upload({ file }, res) {
 
 module.exports = { getUrl, upload };
 
-function readFile(file) {
-  if (file.buffer) return Promise.resolve(file.buffer);
-  return fs.readFile(file.path);
-}
 
-function sha256(...args) {
-  const hash = crypto.createHash('sha256');
-  args.forEach(arg => hash.update(arg));
-  return hash.digest('hex');
-}
